feat(structure): project modularization
